+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Access to "edit injected content" action within language-injected snapshots
   + For example, it allows to preview markdown/mermaid/html ...
 - Structure view
+- Snapshot folding
 
 ## [0.0.1]
 ### Added